Add extension on num to convert to Decimal
To make usage of Decimal more convenient adds an extension on num type
with 'toDecimal()' function that converts current number value to the
Decimal value, this it is possible to write code as:

  final myDecimal = 3.141592.toDecimal();

Of course the `package:decimal/decimal.dart` needs to be imported in
order to use this extension.

Initially I wanted to add an `decimal` getter. But since there are no
getters like that on num type, instead `toInt()` and `toDouble()`
functions are used. Therefore this is also `toDecimal()`.